feat: Authentication, Authorization, Layout

+++ app/Http/Controllers/TaskController.php
@@ -4,13 +4,17 @@ namespace App\Http\Controllers;
 
 use App\Models\Task;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Support\Facades\Validator;
 
 class TaskController extends Controller
 {
     
     public function index(){
-        $tasks = Task::all();
+        //$tasks = Task::all();
+        $tasks = Task::where('user_id', '=', Auth::id())->get();
+
         $txtSrc = '';
         return view('task.index', compact(['tasks', 'txtSrc']));
     }
@@ -32,26 +36,52 @@ class TaskController extends Controller
             return redirect()->route('task.create')->withInput()->withErrors($validator);
         }
 
+        /*
+        $request->request->add(['user_id' => Auth::id()]);
         Task::create($request->all());
+        */
 
+
+        Task::create(
+            [
+                'description' => $request->description,
+                'date' => $request->date,
+                'user_id' => Auth::id(),
+            ]
+        );
+
+        
         return redirect('/task')->with('success', 'Tarefa adicionada com sucesso!');;
     }
 
     public function destroy($id){
 
         $task = Task::findOrFail($id);
+
+        if (!Gate::allows('is-owner', $task)) {
+            //abort(403);
+            return redirect('/task')->with('error', 'Você não pode excluir uma tarefa que não é sua!');
+        }
+
         $task->delete();
 
         return redirect('/task')->with('success', 'Tarefa excluida com sucesso');
     }
 
     public function edit($id){
+
         $task = Task::findOrFail($id);
+
         return view('task.edit', compact('task'));
     }
 
     public function update(Request $request, $id){
         $task = Task::findOrFail($id);
+        
+        if (!Gate::allows('is-owner', $task)) {
+            //abort(403);
+            return redirect('/task')->with('error', 'Você não pode editar uma tarefa que não é sua!');
+        }
 
         $rules = [
             'description' => 'required|min:5',
